requirements.txt
pyaudiowpatch
comtypes
psutil
sounddevice
numpy
soundfile
customtkinter
pillow
pycaw
